+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.32.3 (28.11.2023)
+* Added new method `setSubscriptionAttribute` to support multiple attribute value as array
+
+## 1.32.2 (27.11.2023)
+* Added feature to disable notification permission dialog `setAutoRequestNotificationPermission`
+* Added feature to auto resubscribe `setAutoResubscribe`
+
 ## 1.32.1 (26.11.2023)
 * Fixed version in gradle.properties